Move ldg/quantity macros
