+++ src/corecel/data/detail/LdgIteratorImpl.hh
@@ -13,8 +13,6 @@
 #include "corecel/Types.hh"
 #include "corecel/math/Quantity.hh"
 
-#include "TypeTraits.hh"
-
 namespace celeritas
 {
 namespace detail
