+++ src/corecel/math/detail/QuantityImpl.hh
@@ -6,13 +6,18 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include "corecel/Macros.hh"
+#include <type_traits>
 
-#include "../Constant.hh"
-#include "../NumericLimits.hh"
+#include "corecel/Macros.hh"
+#include "corecel/math/Constant.hh"
+#include "corecel/math/NumericLimits.hh"
 
 namespace celeritas
 {
+//---------------------------------------------------------------------------//
+template<class UnitT, class ValueT>
+class Quantity;
+
 namespace detail
 {
 //---------------------------------------------------------------------------//
@@ -68,6 +73,21 @@ struct AccessorTraits<ResultType (ClassType::*)() const>
 //! Get the result type of a class accessor
 template<class T>
 using AccessorResultType = typename AccessorTraits<T>::result_type;
+
+//---------------------------------------------------------------------------//
+//! Template matching to determine if T is a Quantity
+template<class T>
+struct IsQuantity : std::false_type
+{
+};
+template<class V, class S>
+struct IsQuantity<Quantity<V, S>> : std::true_type
+{
+};
+template<class V, class S>
+struct IsQuantity<Quantity<V, S> const> : std::true_type
+{
+};
 
 //---------------------------------------------------------------------------//
 }  // namespace detail
